requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
tzdata==2025.2
zipp==3.21.0
